Add register.cpp
