+++ Source/lua/register.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <sol/forward.hpp>
+
+#include "lua/lua.hpp"
+
+namespace devilution {
+
+void RegisterAllBindings(sol::state &lua);
+
+} // namespace devilution
